requirements.txt
opencv-python >= 3.4.0
tqdm
packaging >= 20.0
numpy >= 1.17.2
pillow >= 8.3.2
#pytorch >= 1.10
sklearn
scikit-image == 0.19.2
hydra-core == 1.1
matplotlib >= 3.1.2
certifi==2019.11.28
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.3
pyparsing==2.4.6
PyQt5-sip>=4.19.18
PyQt5>=5.12.3
PyQtWebEngine>=5.12.1
python-dateutil>=2.8.1
six==1.14.0
tornado>=6.0.3
